refactor now playing screen, adjust layout spacing
