+++ components/FullSizeMusicPlayer/PlayerControls.tsx
@@ -18,6 +18,7 @@ import {
   useProgress,
 } from "react-native-track-player";
 import { useMusicPlayer } from "../MusicPlayerProvider";
+import { BottomControlBar } from "./BottomControlBar";
 
 interface PlayerControlsProps {
   isSmallScreen: boolean;
@@ -43,7 +44,14 @@ export const PlayerControls = ({
 
   const isPlaying = playbackState === State.Playing;
   return (
-    <View style={{ paddingVertical: isSmallScreen ? 12 : 24 }}>
+    <View
+      style={{
+        paddingTop: isSmallScreen ? 12 : 24,
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "space-between",
+      }}
+    >
       <Slider
         minimumTrackTintColor="white"
         thumbStyle={{ height: 12, width: 12, backgroundColor: "white" }}
@@ -75,39 +83,50 @@ export const PlayerControls = ({
       </View>
       <View
         style={{
-          flexDirection: "row",
-          alignItems: "center",
-          justifyContent: "space-evenly",
-          paddingTop: isSmallScreen ? 8 : 16,
+          flexGrow: 1,
         }}
       >
-        <Pressable onPress={skipToPrevious}>
-          <Ionicons name="play-skip-back-sharp" size={36} color={colors.text} />
-        </Pressable>
-        <PlayPauseTrackButton
-          size={isSmallScreen ? 40 : 60}
-          type={isPlaying ? "pause" : "play"}
-          onPress={togglePlayPause}
-          isEarning={isEarning}
-        />
-        <Pressable onPress={skipToNext}>
-          <Ionicons
-            name="play-skip-forward-sharp"
-            size={36}
-            color={colors.text}
+        <View
+          style={{
+            flexDirection: "row",
+            alignItems: "center",
+            justifyContent: "space-evenly",
+            paddingTop: isSmallScreen ? 8 : 16,
+          }}
+        >
+          <Pressable onPress={skipToPrevious}>
+            <Ionicons
+              name="play-skip-back-sharp"
+              size={36}
+              color={colors.text}
+            />
+          </Pressable>
+          <PlayPauseTrackButton
+            size={isSmallScreen ? 40 : 60}
+            type={isPlaying ? "pause" : "play"}
+            onPress={togglePlayPause}
+            isEarning={isEarning}
           />
-        </Pressable>
+          <Pressable onPress={skipToNext}>
+            <Ionicons
+              name="play-skip-forward-sharp"
+              size={36}
+              color={colors.text}
+            />
+          </Pressable>
+        </View>
+        <Text
+          style={{
+            marginTop: 10,
+            textAlign: "center",
+            fontSize: 12,
+            height: 16,
+          }}
+        >
+          {isEarning ? `Earnings: ${totalEarned} sats` : undefined}
+        </Text>
       </View>
-      <Text
-        style={{
-          marginTop: 10,
-          textAlign: "center",
-          fontSize: 12,
-          height: 16,
-        }}
-      >
-        {isEarning ? `Earnings: ${totalEarned} sats` : undefined}
-      </Text>
+      <BottomControlBar />
     </View>
   );
 };
